[Ruby] Rename red-arrow-gpu to red-arrow-cuda
